configs/zynqmp_kria_kv260_defconfig: fix pmufw sw reset
This patch will solve the problem of the pmufw built by the
zynqmp-pmufw-builder where soft resets crash for the kv260.
Details of the issue can be found here:

https://lore.kernel.org/buildroot/87ilqccu3k.fsf@dell.be.48ers.dk/

Please note that this is a temporary solution as the kv260
pmufw is not yet available on the Xilinx github. This is expected
in the coming months.  The nealfrager github uses the same
licensing terms as the Xilinx github regarding the distribution of
zynqmp pmufw binary files.

https://github.com/Xilinx/ubuntu-firmware/blob/v2022.1_22.04_1/LICENSE
https://github.com/nealfrager/buildroot-firmware/blob/v2022.1/LICENSE

Signed-off-by: Neal Frager <neal.frager@amd.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit 2b564a8ccf8016d4c3c285702b37192abe0852e4)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>