support/testing/tests/package/test_python_flask*: increase time after server startup
It seems like on Gitlab CI, the runners are quite slow, and the Flask
server does not startup in the 15 seconds we give it. So increase this
to 30 seconds before trying to contact the Flask server.

Hopefully fixes:

  https://gitlab.com/buildroot.org/buildroot/-/jobs/3249828594

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 3884cfc0571abc19b35b2f161f6c8c009847a0bb)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>