od: installer: Handle cancelling with SELECT button
Previously, hitting SELECT would cause the dialogs to return code 255.
Since the script would accept any value different than 1 as "accept",
this means that it was possible to flash a system by just pressing
SELECT a few times (or Esc on a keyboard).

Address this issue by handling the error codes properly.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>