configs/nitrogen8*: extend filesystem size
The default ext2 filesystem size of 60 MB is now too small to contain
the root filesystem of the Nitrogen i.MX8 configurations. The
nitrogen8mp_defconfig configuration for examples generates 55 MB of
contents in the rootfs, so an image of 60 MB is slightly too small.

This commit extends the filesystem size to 120 MB.

Fixes:

  https://gitlab.com/buildroot.org/buildroot/-/jobs/3372859635 (nitrogen8mp)
  https://gitlab.com/buildroot.org/buildroot/-/jobs/3372859634 (nitrogen8mn)
  https://gitlab.com/buildroot.org/buildroot/-/jobs/3372859633 (nitrogen8mm)
  https://gitlab.com/buildroot.org/buildroot/-/jobs/3372859632 (nitrogen8m)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>