od: Bump UBIBoot to latest version
Latest version switch Lepus devices to use CL=4 instead of CL=3, which
is a little bit less performant, but is much more reliable and allows to
overclock much more.

It also won't pass memory information, since this has been specified in
the DTS for a while now, and removes mininit 1.x compatibility.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>