board/freescale/imx6ulevk/readme.txt: update broken url
http://freescale.com/ urls are all redirected to the NXP homepage
https://www.nxp.com/

Even if the link is not broken in the sense of a 404 http error, the
pointed resource is not found.

This commit updates the old freescale link to the correct nxp.com
location.

Note: the link now requires a free registration to be accessed.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
(cherry picked from commit 01a3cb0f6eac48150c4bd246402025c0790e2625)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>